# Port Curve25519 encoding operations