--- a/src/ec/curve25519/ops.rs
+++ b/src/ec/curve25519/ops.rs
@@ -14,10 +14,14 @@
 
 //! Elliptic curve operations on Curve25519.
 
+use {bssl, c, error};
 
 // Keep this in sync with `fe` in curve25519/internal.h.
 pub type Elem = [i32; ELEM_LIMBS];
 const ELEM_LIMBS: usize = 10;
+
+// A point on Curve25519, encoded as described in section 5.1.2 of RFC 8032.
+type EncodedPoint = [u8; ELEM_LEN];
 pub const ELEM_LEN: usize = 32;
 
 // Keep this in sync with `ge_p3` in curve25519/internal.h.
@@ -37,6 +41,24 @@ impl ExtPoint {
             z: [0; ELEM_LIMBS],
             t: [0; ELEM_LIMBS],
         }
+    }
+
+    pub fn from_encoded_point_vartime(encoded: &EncodedPoint)
+                          -> Result<Self, error::Unspecified> {
+        let mut point = Self::new_at_infinity();
+
+        try!(bssl::map_result(unsafe {
+            GFp_x25519_ge_frombytes_vartime(&mut point, encoded)
+        }));
+
+        Ok(point)
+    }
+
+    pub fn into_encoded_point(self) -> EncodedPoint {
+        let mut bytes = [0u8; ELEM_LEN];
+        unsafe { GFp_ge_p3_tobytes(&mut bytes, &self); }
+
+        bytes
     }
 
     pub fn invert_vartime(&mut self) {
@@ -63,4 +85,18 @@ impl Point {
             z: [0; ELEM_LIMBS],
         }
     }
+
+    pub fn into_encoded_point(self) -> EncodedPoint {
+        let mut bytes = [0u8; ELEM_LEN];
+        unsafe { GFp_x25519_ge_tobytes(&mut bytes, &self); }
+
+        bytes
+    }
+}
+
+extern {
+    fn GFp_ge_p3_tobytes(s: &mut EncodedPoint, h: &ExtPoint);
+    fn GFp_x25519_ge_frombytes_vartime(h: &mut ExtPoint, s: &EncodedPoint)
+                                       -> c::int;
+    fn GFp_x25519_ge_tobytes(s: &mut EncodedPoint, h: &Point);
 }