In support of #500, we want to oxidize the callers of `fe_invert()`. This work is in support of that intermediate step, as well as inching the Curve25519 arithmetic operations into alignment with the Suite B EC code.