@ranweiler 2017-04-22
At first, for `ExtPoint`, I implemented an `assign_bytes()` method which avoided a `copy_from_slice()` call introduced [here](https://github.com/briansmith/ring/pull/522/commits/7f26e1d560beaf543664c8d3d661d532a83c5e52#diff-3b0b79572e588435f30a2c0a0aa1cff2R148), and instead updated a mutable array reference.

I benchmarked the difference with and without it, and in comparison to upstream master, and the difference seemed negligible, seen here:

```
# Upstream `master`
test signature::ed25519::generate_key_pair                    ... bench:      49,632 ns/iter (+/- 869)
test signature::ed25519::sign_empty                           ... bench:      47,642 ns/iter (+/- 7,259)

# With `assign_bytes()`
test signature::ed25519::generate_key_pair                    ... bench:      49,761 ns/iter (+/- 694)
test signature::ed25519::sign_empty                           ... bench:      47,642 ns/iter (+/- 4,951)

# Without `assign_bytes()`
test signature::ed25519::generate_key_pair                    ... bench:      49,574 ns/iter (+/- 1,020)
test signature::ed25519::sign_empty                           ... bench:      47,670 ns/iter (+/- 4,617)
```