@ranweiler 2017-04-23
Benchmark at 56980ceaebba88315ff4ffcec60bba86ecaf527b: 
```
test signature::ed25519::generate_key_pair                    ... bench:      49,597 ns/iter (+/- 888)
test signature::ed25519::sign_empty                           ... bench:      47,620 ns/iter (+/- 4,508)
```