@ranweiler 2017-04-23
> Please fold the "Use descriptive names for Curve25519 point encoding." change into "Add methods for Curve25519 point encoding."

Done.