@briansmith 2017-04-23
> It looks like we could move self. Grepping a bit, it seems we do that with similar methods in the rsa::bigint code.

> Since it is more idiomatic Rust to prefer refs to self, I would want to keep it as a ref unless we think it would prevent or discourage specific errors.

I don't think refs to self are more (or less) idiomatic in Rust than moving self.

> This seems like a gray area to me. On one hand, an owner of self calling this method multiple times needn't be incorrect per the meaning of the function on its own.

I agree that this is a good rule of thumb for *public* functions, if qualified with "unless there would be a significant performance impact."

> In the actual usage, though, this would always be an error. This method is also private, so we don't have any unforeseen use cases to worry about, so the ref preference is less important.

Exactly.

I don't think it would make any (significant) difference here.  But, in general, in private functions I slightly prefer move over ref because potentially the compiler would have an easier time optimizing the move version. In contrast, move vs. ref makes a big difference in the RSA code because it determines whether we can reuse the inner `BIGNUM` of the moved argument; i.e. using moves helps us avoid mallocs there. So if we're aiming for ultimate consistency across the RSA, Suite B, and Curve25519 code, it probably makes sense to use moves when it's practical and performant to do so, since it is important for the RSA case. (I'm planning to change the Suite B code to be more like the RSA code, when I have time.)

> I'm fine either way. If we do change it, we should change the name from encode() (which is what we ended up with) to into_encoded() or similar. What do you prefer?

I guess `into_encoded_point()` would be most consistent with the RSA code.

I guess if we want to do things the "Idiomatic Rust way" we'll eventually replace all these `into_xxx_xxx` functions with implementations of `Into<XxxXxx>` some day.