@ranweiler 2017-04-23
> But, in general, in private functions I slightly prefer move over ref because potentially the compiler would have an easier time optimizing the move version.

Yes, I suspected that was another reason. Thinking about your other comments about e.g. auto-vectorization, "telling the compiler more" seems preferable, even though I have only a vague sense of what it might get us (I don't feel like I could make a strong concrete argument in this case).

I'll go ahead and rename the methods to `into_encoded_point` and `from_encoded_point` in the commit where they are added.