@ranweiler 2017-04-23
> I'll go ahead and rename the methods to `into_encoded_point` and `from_encoded_point` in the commit where they are added.

Done.