@briansmith 2017-04-24
Thanks! I landed these with minor changes (where my chjanges are seperate commits, except the copyright line change) as:

39b35d3fda039712454ec3bc50362e83ac4c5cae Move Curve25519 operations into their own module.
88579cf0c4e4073515d64abbac9541ec4b49f070 Add methods for Curve25519 point encoding.
49c6a48738a66e1e3bf8a25dc25032940ed2a351 Use array assignment instead of `copy_from_slice()` in one place in ed25519.rs.
6d6aeb6a4fa814b7278565ed2e9036fdcc962edb Port Curve25519 point encoding to Rust.
bc5bd4a0bdd151c3cab2e33b797efa65333c5fb1 Clarify `unsafe` and constant-timedness of Curve25519 `encode_point()`.
7d46aad3e2aaad9cd6f5530fa34c586f9914716f Move all non-protocol Curve25519 types into `ops` module.

Please take some time to review my changes, which were all minor (and hopefully correct).