@briansmith 2017-04-24
BTW, I think this is all really good. The added clarity from the logical refactoring you've done has made things a lot clearer.